requirements.txt
python-dotenv>=1.0.0
requests>=2.28.0
streamlit>=1.30.0
pandas>=2.0.0

# Development dependencies
pytest>=7.0.0
pytest-cov>=4.0.0